\docType{data}
\name{pvRome}
\alias{pvRome}
\title{Sun irradiance and air temperature in Rome}
\description{
  Sun irradiance and air temperature in Rome from October
  2012 to November 2012 as scraped daily from
  \url{meteo.enel.it}.
}
\author{
  Marco Bascietto \email{marco@bascietto.name}
}
\keyword{data}
\keyword{datasets}

