\name{trim}
\alias{trim}
\title{Trim leading and trailing spaces from a string}
\usage{
  trim(x)
}
\arguments{
  \item{x}{the string to be cleaned}
}
\value{
  a string with leading and trailing spaces removed
}
\description{
  Trim leading and trailing spaces from a string
}
\author{
  Marco Bascietto \email{marco@bascietto.name}
}

