% Generated by roxygen2 (4.0.0): do not edit by hand
\name{getSunCoordinates}
\alias{getSunCoordinates}
\title{Estimates Sun azimuth and height above a location at a given time}
\usage{
getSunCoordinates(locCoordinates = c(6.5, 46.5), time = Sys.time())
}
\arguments{
\item{locCoordinates}{a vector holding longitude and latitude of a single location on earth}

\item{time}{a POSIXct class object}
}
\value{
a matrix of 2 columns, \code{azimuth} and \code{height} of sun
}
\description{
Estimates Sun azimuth and height above a location at a given time
}
\author{
Marco Bascietto \email{marco@bascietto.name}
}
\references{
Code adapted from \url{http://stackoverflow.com/questions/8708048/position-of-the-sun-given-time-of-day-latitude-and-longitude}, Richie Cotton answer
}

