% Generated by roxygen2 (4.0.0): do not edit by hand
\docType{package}
\name{meteopv-package}
\alias{meteopv}
\alias{meteopv-package}
\title{Scraping from the Web the hourly solar irradicance and air temperature to estimate efficiency and energy produced by photovoltaic cell systems.}
\description{
A simpe example of meteorological data elaboration is given in \code{\link{pvRome}} dataset.
}
\author{
Marco Bascietto \email{marco@bascietto.name}
}
\references{
Source code is hosted at GitHub (\url{http://mbask.github.com/meteopv/})
}
\keyword{package}

